Photo-sorting/requirements.txt
streamlit==1.22.0
tensorflow==2.12.0
numpy==1.24.3
Pillow==9.5.0
requests==2.31.0
matplotlib==3.7.1